hello_world/requirements.txt
requests
wheel
mecab-python3==1.0.4
